Fix edgeql-go file write race condition (#395)
To find the package name, edgeql-go reads the other go files in the
directory. This was happening at the same time that generated files in
that directory were being written. If a file being written in one go
routine was read in another it would not parse because it was
incomplete. This caused edgeql-go to crash.

The package name is now determined only once for a directory before
processing any of the queries in that directory.